project hero page color updated
