+++ src/app/projects/page.tsx
@@ -2,6 +2,7 @@ import { ProjectsPage } from "@/components/custom/projects";
 import React from "react";
 import { ProjectCard } from "./card";
 import Hero from "./hero";
+import Navigation from "./navigation";
 const projects = [
   {
     id: 1,
@@ -69,6 +70,7 @@ const page = () => {
   return (
     <div className="">
       <div className="h-20 md:h-10"></div>
+   <Navigation/>
    <Hero/>
    <h3 className="text-3xl ml-10 font-bold ">Popular Projects </h3>
    <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8 mx-10 md:mx-10 my-10">
